+++ src/game/TextBox.hpp
@@ -21,7 +21,7 @@ public:
 
 	// Game functions
 	void draw(sf::RenderWindow* window);
-	void update(sf::Vector2i mouseLoc, sf::Event event);
+	void update(sf::Vector2f mouseLoc, sf::Event event);
 
 	// Assessors
 	void setPosition( float x, float y);
